refactor(studio): separate search context + provider to work around HMR issue with vite/plugin-react
