+++ packages/sanity/src/core/studio/components/navbar/__workshop__/SearchDialogStory.tsx
@@ -1,7 +1,7 @@
 import {LayerProvider} from '@sanity/ui'
 import React from 'react'
 import {SearchDialog} from '../search'
-import {SearchProvider} from '../search/contexts/search'
+import {SearchProvider} from '../search/contexts/search/SearchProvider'
 
 const noop = () => null
