+++ packages/sanity/src/core/studio/components/navbar/search/SearchDialog.tsx
@@ -8,7 +8,7 @@ import {SearchHeader} from './components/SearchHeader'
 import {SearchResults} from './components/SearchResults'
 import {TypeFilters} from './components/TypeFilters'
 import {CommandListProvider} from './contexts/commandList'
-import {useSearchState} from './contexts/search'
+import {useSearchState} from './contexts/search/useSearchState'
 import {useMeasureSearchResultsIndex} from './hooks/useMeasureSearchResultsIndex'
 import {useSearchHotkeys} from './hooks/useSearchHotkeys'
 import {hasSearchableTerms} from './utils/hasSearchableTerms'
