+++ packages/sanity/src/core/studio/components/navbar/search/components/PlaceholderSearchInput.tsx
@@ -4,7 +4,7 @@ import {Flex, KBD, TextInput} from '@sanity/ui'
 import React, {forwardRef, KeyboardEvent as ReactKeyboardEvent, Ref, useCallback} from 'react'
 import styled from 'styled-components'
 import {GLOBAL_SEARCH_KEY, GLOBAL_SEARCH_KEY_MODIFIER} from '../constants'
-import {useSearchState} from '../contexts/search'
+import {useSearchState} from '../contexts/search/useSearchState'
 
 const KeyboardShortcutFlex = styled(Flex)`
   position: absolute;
