+++ packages/sanity/src/core/studio/components/navbar/search/components/SearchPopover.tsx
@@ -5,7 +5,7 @@ import styled, {css} from 'styled-components'
 import {useColorScheme} from '../../../../colorScheme'
 import {POPOVER_INPUT_PADDING, POPOVER_MAX_HEIGHT, POPOVER_MAX_WIDTH} from '../constants'
 import {CommandListProvider} from '../contexts/commandList'
-import {useSearchState} from '../contexts/search'
+import {useSearchState} from '../contexts/search/useSearchState'
 import {useMeasureSearchResultsIndex} from '../hooks/useMeasureSearchResultsIndex'
 import {useSearchHotkeys} from '../hooks/useSearchHotkeys'
 import {hasSearchableTerms} from '../utils/hasSearchableTerms'
