+++ packages/sanity/src/core/studio/components/navbar/search/components/SearchResults.tsx
@@ -5,7 +5,7 @@ import styled from 'styled-components'
 import {getPublishedId} from '../../../../../util/draftUtils'
 import {VIRTUAL_LIST_ITEM_HEIGHT, VIRTUAL_LIST_OVERSCAN} from '../constants'
 import {useCommandList} from '../contexts/commandList'
-import {useSearchState} from '../contexts/search'
+import {useSearchState} from '../contexts/search/useSearchState'
 import {NoResults} from './NoResults'
 import {PointerOverlay} from './PointerOverlay'
 import {SearchError} from './SearchError'
