+++ packages/sanity/src/core/studio/components/navbar/search/components/SortMenu.tsx
@@ -15,7 +15,7 @@ import isEqual from 'lodash/isEqual'
 import React, {useCallback, useId, useMemo} from 'react'
 import styled from 'styled-components'
 import {SUBHEADER_HEIGHT_LARGE, SUBHEADER_HEIGHT_SMALL} from '../constants'
-import {useSearchState} from '../contexts/search'
+import {useSearchState} from '../contexts/search/useSearchState'
 import {
   SearchOrdering,
   ORDER_CREATED_ASC,
