+++ packages/sanity/src/core/studio/components/navbar/search/components/TypeFilterItem.tsx
@@ -4,7 +4,7 @@ import React, {useCallback} from 'react'
 import styled, {css} from 'styled-components'
 import type {SearchableType} from '../../../../../search'
 import {useCommandList} from '../contexts/commandList'
-import {useSearchState} from '../contexts/search'
+import {useSearchState} from '../contexts/search/useSearchState'
 
 interface TypeFilterItemProps {
   index: number
