+++ packages/sanity/src/core/studio/components/navbar/search/components/TypeFilters.tsx
@@ -5,7 +5,7 @@ import styled from 'styled-components'
 import {useSchema} from '../../../../../hooks'
 import {SUBHEADER_HEIGHT_LARGE, SUBHEADER_HEIGHT_SMALL} from '../constants'
 import {CommandListProvider} from '../contexts/commandList'
-import {useSearchState} from '../contexts/search'
+import {useSearchState} from '../contexts/search/useSearchState'
 import {getSelectableOmnisearchTypes} from '../utils/selectors'
 import {supportsTouch} from '../utils/supportsTouch'
 import {CustomTextInput} from './CustomTextInput'
